requirements.txt
tensorflow>=2.9.0
numpy>=1.18.0
joblib>=1.2.0
scikit-fuzzy>=0.4.2
pyserial>=3.5
